Remove Trainable (#8)
Remove Trainable trait from tree_trainer. In new approach we don't need a callback. During tree training, the targets array (in Trainset) is rearranged according to the tree structure. We return it with information about incidence of ranges to tree nodes, so that the caller can do processing of leaves.

* Rename tree_trainer into tree_builder
* Move Trainset to tree_builder
* Add Splittable trait to decision_tree module
* Remove Trainable trait from tree_builder