Compact tree (#15)
In new tree we store only internal nodes. Each node has 2 u32
pointers to left and right child. If children are leaves, then a u32
pointer is treated as leaf value. For classification tasks, this
value is an index of probabilities group. For regression tasks, u32
is converted into f32.